requirements.txt
llama-index-core==0.11.17